+++ src/main/java/org/example/entities/Address.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.entities;
 
 import javax.persistence.*;
