+++ src/main/java/org/example/entities/Customer.java
@@ -1,9 +1,11 @@
-package org.example;
+package org.example.entities;
+
+import org.example.entities.Address;
+import org.example.entities.Concert;
 
 import javax.persistence.*;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 //import java.util.Date;
