+++ src/main/java/org/example/enums/Setting.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.enums;
 
 import javax.persistence.*;
