+++ src/main/resources/hibernate.cfg.xml
@@ -16,10 +16,10 @@
         <property name="format_sql">false</property>
         <property name="hbm2ddl.auto">update</property>
 
-        <mapping class="org.example.Address"/>
-        <mapping class="org.example.Arena"/>
-        <mapping class="org.example.Concert"/>
-        <mapping class="org.example.Customer"/>
+        <mapping class="org.example.entities.Address"/>
+        <mapping class="org.example.entities.Arena"/>
+        <mapping class="org.example.entities.Concert"/>
+        <mapping class="org.example.entities.Customer"/>
 <!--        <mapping class="org.example.Wc"/>-->
     </session-factory>
 </hibernate-configuration>
